fix: common 모듈에서 spring 관련 의존성 제거, 에러 핸들러 및 공통 에러 응답 제거
