--- common/src/main/kotlin/com/hunhui/common/error/ErrorCode.kt
@@ -1,9 +1,0 @@
-package com.hunhui.common.error
-
-import org.springframework.http.HttpStatus
-
-interface ErrorCode {
-    val code: String
-    val httpStatus: HttpStatus
-    val message: String
-}
