+++ common/src/main/kotlin/com/hunhui/ticketworld/common/error/ErrorCode.kt
@@ -1,0 +1,6 @@
+package com.hunhui.ticketworld.common.error
+
+interface ErrorCode {
+    val code: String
+    val message: String
+}
