+++ common/src/main/kotlin/com/hunhui/ticketworld/common/vo/Money.kt
@@ -1,0 +1,9 @@
+package com.hunhui.ticketworld.common.vo
+
+data class Money(
+    val amount: Long,
+) {
+    init {
+        require(amount >= 0) { "Amount must be greater than or equal to 0" }
+    }
+}
