--- common/src/main/resources/application.yml
@@ -1,3 +1,0 @@
-spring:
-    application:
-        name: common
