+++ common/src/test/kotlin/com/hunhui/ticketworld/common/CommonTests.kt
@@ -1,0 +1,9 @@
+package com.hunhui.ticketworld.common
+
+import org.junit.jupiter.api.Test
+
+class CommonTests {
+    @Test
+    fun contextLoads() {
+    }
+}
